pyproject.toml
[project]
name = "repo-agent-cli"
version = "0.1.0"
description = "CLI agent over your codebase + CI/CD"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pydantic>=2.7",
  "openai>=1.0.0",
  "numpy>=1.26",
  "tqdm>=4.66",
  "pyyaml>=6.0",
  "gitpython>=3.1"
]
[project.scripts]
agent = "agent_cli.main:app"
